First stab at fixing cached_repository operation
It seems like CachedRepository.normalize_rev expects rev IDs to be
integers, which obviously isn't going to work here.  This version
overrides normalize_rev to use the method from git_fs.  I must admit I'm
not entirely sure whether this is the best idea, but it seems to make
sense at the moment :)